+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [CalVer](https://calver.org/).
 
-## [2025.3.2] - 2025-28-3
+## [2025.40] - 2025-28-3
+
+- Danish search with `sortByScore` is improved so more relevant hits get ranked higher. `sortByScore` is now default.
 
 ### Added